update opcode testing to Cancun
