+++ contracts/hardhat_overrides.ts
@@ -17,9 +17,6 @@ const londonOverridePaths = [
   "src/tokens/LineaVoyageXP.sol",
   "src/tokens/LineaSurgeXP.sol",
   "src/tokens/TokenMintingRateLimiter.sol",
-  "src/_testing/unit/opcodes/ErrorAndDestructionTesting.sol",
-  "src/_testing/unit/opcodes/OpcodeTestContract.sol",
-  "src/_testing/unit/opcodes/OpcodeTester.sol",
   "src/_testing/mocks/bridging/MockMessageService.sol",
   "src/_testing/mocks/bridging/MockMessageServiceV2.sol",
   "src/_testing/mocks/bridging/MockTokenBridge.sol",
