--- contracts/local-deployments-artifacts/static-artifacts/LondonEvmCodes.json
@@ -1,12 +1,0 @@
-{
-  "contractName": "LondonEvmCodes",
-  "sourceName": "contracts/test-contracts/LondonEvmCodes.yul",
-  "abi": [
-    {
-      "payable": true,
-      "stateMutability": "payable",
-      "type": "fallback"
-    }
-  ],
-  "bytecode": "0x61047c61001060003961047c6000f3fe610007610452565b63a378ff3e811461001757600080fd5b61001f61002d565b610027610092565b5061047b565b63fed44325604051818152600080600483600060555af18061004e57600080fd5b600080600484600060555af290508061006657600080fd5b60008060048460555af490508061007c57600080fd5b6004820191506712345678123456788253505050565b602060206001601f600263fffffffffa5060006000600042f550600060006000f050600060006000600060006000a460006000600060006000a36000600060006000a2600060006000a160006000a0585059505a50426000556000545060004050415042504350445045504650475048507300000000000000000000000000000000000000003f506000600060003e6000600060007300000000000000000000000000000000000000003c3d507300000000000000000000000000000000000000003b5060016001016001026003036001046001056001066001076001600108600160010960020160030a60010b600810600a11600112600113600114156001166001176001181960161a60011b60011c60011d506000600020303132333450505050503635600060003738604051600081016000600083393a50505050607e50617e0150627e012350637e01234550647e0123456750657e012345678950667e0123456789ab50677e0123456789abcd50687e0123456789abcdef50697e0123456789abcdef01506a7e0123456789abcdef0123506b7e0123456789abcdef012345506c7e0123456789abcdef01234567506d7e0123456789abcdef0123456789506e7e0123456789abcdef0123456789ab506f7e0123456789abcdef0123456789abcd50707e0123456789abcdef0123456789abcdef50717e0123456789abcdef0123456789abcdef0150727e0123456789abcdef0123456789abcdef012350737e0123456789abcdef0123456789abcdef01234550747e0123456789abcdef0123456789abcdef0123456750757e0123456789abcdef0123456789abcdef012345678950767e0123456789abcdef0123456789abcdef0123456789ab50777e0123456789abcdef0123456789abcdef0123456789abcd50787e0123456789abcdef0123456789abcdef0123456789abcdef50797e0123456789abcdef0123456789abcdef0123456789abcdef01507a7e0123456789abcdef0123456789abcdef0123456789abcdef0123507b7e0123456789abcdef0123456789abcdef0123456789abcdef012345507c7e0123456789abcdef0123456789abcdef0123456789abcdef01234567507d7e0123456789abcdef0123456789abcdef0123456789abcdef0123456789507e0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcd507f0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef808182838485868788898a8b8c8d8e8f909192939495969798999a9b9c9d9e9f5050505050505050505050505050505050565b60007c010000000000000000000000000000000000000000000000000000000060003504905090565b"
-}
