+++ contracts/local-deployments-artifacts/static-artifacts/ShanghaiEvmCodes.json
@@ -1,0 +1,12 @@
+{
+  "contractName": "ShanghaiEvmCodes",
+  "sourceName": "contracts/test-contracts/ShanghaiEvmCodes.yul",
+  "abi": [
+    {
+      "payable": true,
+      "stateMutability": "payable",
+      "type": "fallback"
+    }
+  ],
+  "bytecode": "0x61047361000e5f396104735ff3fe61000761044b565b63a378ff3e8114610016575f5ffd5b61001e61002c565b610026610089565b50610472565b63fed443256040518181525f5f6004835f60555af18061004a575f5ffd5b5f5f6004845f60555af290508061005f575f5ffd5b5f5f60048460555af4905080610073575f5ffd5b6004820191506712345678123456788253505050565b5f50602060206001601f600263fffffffffa5060006000600042f550600060006000f050600060006000600060006000a460006000600060006000a36000600060006000a2600060006000a160006000a0585059505a50426000556000545060004050415042504350445045504650475048507300000000000000000000000000000000000000003f506000600060003e6000600060007300000000000000000000000000000000000000003c3d507300000000000000000000000000000000000000003b5060016001016001026003036001046001056001066001076001600108600160010960020160030a60010b600810600a11600112600113600114156001166001176001181960161a60011b60011c60011d506000600020303132333450505050503635600060003738604051600081016000600083393a50505050607e50617e0150627e012350637e01234550647e0123456750657e012345678950667e0123456789ab50677e0123456789abcd50687e0123456789abcdef50697e0123456789abcdef01506a7e0123456789abcdef0123506b7e0123456789abcdef012345506c7e0123456789abcdef01234567506d7e0123456789abcdef0123456789506e7e0123456789abcdef0123456789ab506f7e0123456789abcdef0123456789abcd50707e0123456789abcdef0123456789abcdef50717e0123456789abcdef0123456789abcdef0150727e0123456789abcdef0123456789abcdef012350737e0123456789abcdef0123456789abcdef01234550747e0123456789abcdef0123456789abcdef0123456750757e0123456789abcdef0123456789abcdef012345678950767e0123456789abcdef0123456789abcdef0123456789ab50777e0123456789abcdef0123456789abcdef0123456789abcd50787e0123456789abcdef0123456789abcdef0123456789abcdef50797e0123456789abcdef0123456789abcdef0123456789abcdef01507a7e0123456789abcdef0123456789abcdef0123456789abcdef0123507b7e0123456789abcdef0123456789abcdef0123456789abcdef012345507c7e0123456789abcdef0123456789abcdef0123456789abcdef01234567507d7e0123456789abcdef0123456789abcdef0123456789abcdef0123456789507e0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcd507f0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef808182838485868788898a8b8c8d8e8f909192939495969798999a9b9c9d9e9f5050505050505050505050505050505050565b5f7c01000000000000000000000000000000000000000000000000000000005f3504905090565b"
+}
